+++ src/extensions/renderer/canvas/webgl/defaults.js
@@ -1,0 +1,17 @@
+import { defaults } from '../../../../util';
+
+export const RENDER_TARGET = {
+  SCREEN:  { name: 'screen',  screen:  true },
+  PICKING: { name: 'picking', picking: true },
+};
+
+export const renderDefaults = defaults({
+  getKey: null,
+  drawElement: null,
+  getBoundingBox: null,
+  getRotation: null,
+  getRotationPoint: null,
+  getRotationOffset: null,
+  isVisible: null,
+  getPadding: null,
+});
